refactor: organize library files
