+++ package.json
@@ -15,8 +15,8 @@
   "license": "Unlicense",
   "author": "Alfi Maulana <alfi.maulana.f@gmail.com>",
   "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/lib.js",
+  "types": "dist/lib.d.ts",
   "bin": "dist/bin.js",
   "files": [
     "dist"
